Samples occurrences form for Splash project which has custom submission code.
Note that editing of existing data on the species grid is not currently supported.

The format is as follows,
Two grids: Trees and Epiphytes
For each tree, a sub-sample is created as it needs its own grid reference. 
Each tree also has a tree occurrence.
Each tree can also have any number of Epiphyte occurrences.
The Tree and Epiphyte occurrences are held as submodels of the tree sub-sample.

On screen, the Epiphyte occurrences are held on a seperate grid to the trees.
There is an Epiphyte per row. 
Each column is a count of Epiphytes for trees 1 to 10 held as occurrence attributes.
If a particular Epiphyte has a count associated with it for a given tree, then we know to create an occurrence for 
that Epiphyte.