Fix failures connecting to NBN
Failed with errors
 - SSL operation failed with code 1.
 - Failed to enable crypto
 - Failed to open stream.
Doubtful this is the best fix as there is now a certificate file to
maintain in the module. Wondering if there could have been some fix at
the server level. Don't know what changed that meant this error started
occurring when it had been okay. Potentially some change at the NBN end.